Moved test helpers to tests/helpers and added global path to them.
