+++ jest.config.js
@@ -14,6 +14,7 @@ module.exports = {
         "math/(.*)$": "<rootDir>/app/core/ts/utils/math/$1.ts",
         "core/(.*)$": "<rootDir>/app/core/ts/$1.ts",
         "digital/(.*)$": "<rootDir>/app/digital/ts/$1.ts",
-        "analog/(.*)$": "<rootDir>/app/analog/ts/$1.ts"
+        "analog/(.*)$": "<rootDir>/app/analog/ts/$1.ts",
+        "test/helpers/(.*)": "<rootDir>/tests/helpers/$1.ts"
     }
 };
