+++ tests/app/digital/ts/models/ioobjects/flipflops/SRFlipFlop.test.ts
@@ -5,7 +5,7 @@ import {Switch}          from "digital/models/ioobjects/inputs/Switch";
 import {SRFlipFlop}      from "digital/models/ioobjects/flipflops/SRFlipFlop";
 import {LED}             from "digital/models/ioobjects/outputs/LED";
 
-import {Place, Connect} from "../../../Helpers";
+import {Place, Connect} from "test/helpers/Helpers";
 
 describe("SRFlipFLop", () => {
     const designer = new DigitalCircuitDesigner(0);
