+++ tests/app/digital/ts/models/ioobjects/gates/ANDGate.test.ts
@@ -5,7 +5,7 @@ import {Switch}          from "digital/models/ioobjects/inputs/Switch";
 import {ANDGate}         from "digital/models/ioobjects/gates/ANDGate";
 import {LED}             from "digital/models/ioobjects/outputs/LED";
 
-import {Place, Connect} from "../../../Helpers";
+import {Place, Connect} from "test/helpers/Helpers";
 
 describe("ANDGate", () => {
     describe("ANDGate", () => {
