+++ tests/app/digital/ts/models/ioobjects/gates/ORGate.test.ts
@@ -5,7 +5,7 @@ import {Switch}          from "digital/models/ioobjects/inputs/Switch";
 import {ORGate}          from "digital/models/ioobjects/gates/ORGate";
 import {LED}             from "digital/models/ioobjects/outputs/LED";
 
-import {Place, Connect} from "../../../Helpers";
+import {Place, Connect} from "test/helpers/Helpers";
 
 describe("ORGate", () => {
     describe("ORGate", () => {
