+++ tests/app/digital/ts/models/ioobjects/inputs/Button.test.ts
@@ -6,7 +6,7 @@ import {Button}          from "digital/models/ioobjects/inputs/Button";
 import {ANDGate}         from "digital/models/ioobjects/gates/ANDGate";
 import {LED}             from "digital/models/ioobjects/outputs/LED";
 
-import {Place, Connect} from "../../../Helpers";
+import {Place, Connect} from "test/helpers/Helpers";
 
 describe("Button", () => {
     const designer = new DigitalCircuitDesigner(0);
