+++ tests/app/digital/ts/models/ioobjects/other/IC.test.ts
@@ -9,7 +9,7 @@ import {LED}             from "digital/models/ioobjects/outputs/LED";
 import {IC}     from "digital/models/ioobjects/other/IC";
 import {ICData} from "digital/models/ioobjects/other/ICData";
 
-import {Place, Connect} from "../../../Helpers";
+import {Place, Connect} from "test/helpers/Helpers";
 
 describe("IC", () => {
     test("Basic IC", () => {
