+++ tests/app/digital/ts/utils/tools/RotateTool.test.ts
@@ -14,9 +14,9 @@ import {ANDGate}         from "digital/models/ioobjects/gates/ANDGate";
 import {ORGate}          from "digital/models/ioobjects/gates/ORGate";
 
 import {FakeInput} from "../FakeInput";
-import {InitializeInput, CreateDefaultToolManager} from "./Helpers";
+import {InitializeInput, CreateDefaultToolManager} from "test/helpers/ToolHelpers";
 
-import {Place} from "../../Helpers";
+import {Place} from "test/helpers/Helpers";
 
 describe("Rotate Tool", () => {
     const camera = new Camera(500, 500);
