+++ tests/app/digital/ts/utils/tools/TranslateTool.test.ts
@@ -14,9 +14,9 @@ import {LED}         from "digital/models/ioobjects/outputs/LED";
 import {DigitalNode} from "digital/models/ioobjects/other/DigitalNode";
 
 import {FakeInput} from "../FakeInput";
-import {InitializeInput, CreateDefaultToolManager} from "./Helpers";
+import {InitializeInput, CreateDefaultToolManager} from "test/helpers/ToolHelpers";
 
-import {Place, Connect} from "../../Helpers";
+import {Place, Connect} from "test/helpers/Helpers";
 
 describe("Translate Tool", () => {
     const camera = new Camera(500, 500);
